3_crew/prism/pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "PRISM - Position Risk Intelligence & Swap Monitor"
requires-python = ">=3.11,<3.13"

dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.12.0",
    "psycopg2-binary>=2.9.0", # database connection
    "python-dotenv>=1.0.0", # environment variables
]

[project.scripts]
run_crew = "prism.crew:main"

[tool.uv]
package = true
